move http calls into separate wrapper class
