+++ src/WebApi/End2endTests/QuestionTests/GetQuestionTests.cs
@@ -30,7 +30,7 @@ public class GetQuestionTests : E2ETestsBase, IClassFixture<MyWebApplicationFact
             queryParams.Add("difficulties", difficulty.ToString());
         }
 
-        return await GetAsync<IEnumerable<QuestionModel>>(_url, queryParams);
+        return await _wrapper.GetAsync<IEnumerable<QuestionModel>>(_url, queryParams);
     }
 
     [Fact]
